[refs #15] Add run.sh to build and launch.
The launch script run.sh will run docker-compose. docker-compose build - build a restshorturl:dev image. docker-compose up start the latest mysql and restshorturl:dev. In testing the mysql container often initializes after restshorturl, this highlighted that the app was failing to start if the db was not ready. This has been changed in application-prod.properties with continue-on-error=true. Also now using schema.sql to load the schema rather than hibernate audo-ddl as this also causes app initialization to fail if database is not ready. Log to console and file in prod profile.